Lazy load react-virtualized
